WIP: moving functions into their respective modules
