+++ src/neutralb1/utils.py
@@ -3,8 +3,10 @@
 import os
 import pathlib
 import re
-import subprocess
 import sys
+from typing import Dict, List
+
+import pandas as pd
 
 
 def sort_input_files(input_files: list, position: int = -1) -> list:
@@ -85,3 +87,127 @@ def get_workspace_dir() -> str:
         raise FileNotFoundError("Could not find 'neutralb1' directory in the path.")
 
     return workspace_dir
+
+
+def parse_amplitude(amp: str) -> Dict[str, str]:
+    """
+    Parse 'eJPmL' style amplitude string into its individual quantum numbers.
+
+    Extracts reflectivity (e), total angular momentum (J), parity (P), m-projection (m),
+    and orbital angular momentum (L) from amplitude strings like 'p1p0S'. Can also
+    handle coherent sum variations like "m1m" or "p".
+
+    Args:
+        amp (str): Amplitude string, e.g. 'p1p0S', 'm1mpP', etc.
+
+    Returns:
+        dict: Dictionary with keys 'e', 'J', 'P', 'm', 'L'. Values are extracted from
+        the input string, or "" if not found.
+    """
+
+    result_dict = {"e": "", "J": "", "P": "", "m": "", "L": ""}
+    regex_dict = {
+        "e": r"(?<![0-9]{1}[pm]{1})(?<!\d)[pm]",
+        "JP": r"[0-9]{1}[pm]{1}",  # j and p always appear together
+        "m": r"([0-9]{1}[pm]{1})+[lnm0pqr]",  # assumes "m" is always proceeded by "JP"
+        "L": r"[SPDF]",
+    }
+
+    for quantum_number, expression in regex_dict.items():
+        search = re.search(expression, amp)
+        if search:
+            if quantum_number == "m":
+                # the search actually returns "JPm", so grab the last char if found
+                result_dict["m"] = search.group()[-1]
+            elif quantum_number == "JP":
+                result_dict["J"] = search.group()[0]
+                result_dict["P"] = search.group()[1]
+            else:
+                result_dict[quantum_number] = search.group()
+
+    return result_dict
+
+
+def get_coherent_sums(df: pd.DataFrame) -> Dict[str, List[str]]:
+    """Returns a dict of coherent sums from a fit results dataframe
+
+    Args:
+        df (pd.DataFrame): dataframe of fit results loaded from csv
+    Returns:
+        dict: Is of the form {Coherent sum string: set(amplitudes)}
+        e.g. {"eJPmL": ["p1p0S", "m1mpP", ...]}
+    """
+    # create an empty list for every type of coherent sum
+    sum_types = ["eJPmL", "JPmL", "eJPL", "JPL", "eJP", "JP", "e"]
+    coherent_sums = {d: set() for d in sum_types}
+
+    # grab all eJPml columns
+    for column in df.columns:
+        # skip the phase difference columns and any status columns
+        if "_" in column or len(column) > 5:
+            continue
+
+        result_dict = parse_amplitude(column)
+        # only add to key if all elements of key are in the column
+        for key in coherent_sums.keys():
+            split_key = list(key.lower())
+            if any(result_dict[char] == "" for char in split_key):
+                continue
+            coh_sum = "".join([result_dict[char] for char in split_key])
+            coherent_sums[key].add(coh_sum)
+
+    return {k: sorted(v) for k, v in coherent_sums.items()}  # sort each set
+
+
+def convert_amp_name(input_string: str) -> str:
+    """Converts amplitude string to J^P L_m^(e) LaTeX style string
+
+    Function can handle both amplitudes and phase differences. If input_string is not
+    of eJPmL format, or subset of it i.e. eJPL, then the output will be undefined.
+
+    Args:
+        input_string (str): string in eJPmL format
+    Returns:
+        str: Prettier LaTeX style string in J^P L_m^(e) format. If it's a phase
+            difference it's then J^P L_m^(e) - J^P L_m^(e)
+    """
+
+    pm_dict = {"m": "-", "p": "+", "": ""}
+    m_proj_dict = {"n": -2, "m": -1, "0": 0, "p": +1, "q": +2, "": ""}
+
+    # CASE 1: phase difference, always of form 'eJPmL_eJPmL'
+    if "_" in input_string:
+        amps = input_string.split("_")
+        if len(amps) != 2:
+            raise ValueError("Phase difference must be in 'eJPmL_eJPmL' format!")
+        e1, j1, p1, m1, l1 = (
+            pm_dict[amps[0][0]],
+            amps[0][1],
+            pm_dict[amps[0][2]],
+            m_proj_dict[amps[0][3]],
+            amps[0][-1],
+        )
+        e2, j2, p2, m2, l2 = (
+            pm_dict[amps[1][0]],
+            amps[1][1],
+            pm_dict[amps[1][2]],
+            m_proj_dict[amps[1][3]],
+            amps[1][-1],
+        )
+
+        return (
+            rf"${j1}^{{{p1}}}{l1}_{{{m1}}}^{{({e1})}}"
+            rf" - {j2}^{{{p2}}}{l2}_{{{m2}}}^{{({e2})}}$"
+        )
+
+    # CASE 2: typical amplitude coherent sum
+    amp_dict = parse_amplitude(input_string)
+
+    # set each quantum number to its found value. If not found, denote it with a sum
+    e = r"\Sigma\varepsilon" if amp_dict["e"] == "" else pm_dict[amp_dict["e"]]
+    j = r"\Sigma J" if amp_dict["j"] == "" else amp_dict["j"]
+    p = "" if amp_dict["p"] == "" else pm_dict[amp_dict["p"]]
+    m = r"\Sigma m" if amp_dict["m"] == "" else m_proj_dict[amp_dict["m"]]
+    l = r"\Sigma \ell" if amp_dict["l"] == "" else amp_dict["l"]
+
+    return rf"${j}^{{{p}}}{l}_{{{m}}}^{{({e})}}$"
